Implement PR feedback: Rename access class to AccessEsm1p5, add time constants, and reduce test file size
