+++ src/model_config_tests/test_bit_reproducibility.py
@@ -10,6 +10,7 @@ from typing import Optional
 import pytest
 
 from model_config_tests.exp_test_helper import setup_exp
+from model_config_tests.util import SECONDS_IN_DAY, SECONDS_IN_HOUR
 
 
 def set_checksum_output_dir(output_path: Path):
@@ -103,7 +104,7 @@ class TestBitReproducibility:
         )
 
         # Set the checksum output filename using the model default runtime
-        runtime_hours = exp.model.default_runtime_seconds // 3600
+        runtime_hours = exp.model.default_runtime_seconds // SECONDS_IN_HOUR
         checksum_filename = f"historical-{runtime_hours}hr-checksum.json"
 
         # Read the historical checksum file
@@ -162,7 +163,7 @@ class TestBitReproducibility:
         exp_2x1day = setup_exp(control_path, output_path, "test_restart_repro_2x1day")
 
         # Reconfigure to a 1 day run.
-        exp_2x1day.model.set_model_runtime(seconds=86400)
+        exp_2x1day.model.set_model_runtime(seconds=SECONDS_IN_DAY)
 
         # Now run twice.
         exp_2x1day.setup_and_run()
@@ -171,7 +172,7 @@ class TestBitReproducibility:
         # Now do a single 2 day run
         exp_2day = setup_exp(control_path, output_path, "test_restart_repro_2day")
         # Reconfigure
-        exp_2day.model.set_model_runtime(seconds=172800)
+        exp_2day.model.set_model_runtime(seconds=(2 * SECONDS_IN_DAY))
 
         # Run once.
         exp_2day.setup_and_run()
